Refactor TheFeed Move feed and following logic to useFeedPosts and useFollowing hooks Move PostDto to it's own file
