+++ frontend/src/types/PostDto.ts
@@ -1,0 +1,9 @@
+export type PostDto = {
+  id: string;
+  senderId: string;
+  senderUsername: string;
+  receiverId: string;
+  receiverUsername: string;
+  content: string;
+  createdAt: string;
+};
